Undo fsdp for unet

+++ ldm/modules/diffusionmodules/util.py
@@ -240,8 +240,7 @@ class SiLU(nn.Module):
 class GroupNorm32(nn.GroupNorm):
 
     def forward(self, x):
-        # return super().forward(x.float()).type(x.dtype)
-        return super().forward(x)
+        return super().forward(x.float()).type(x.dtype)
 
 def conv_nd(dims, *args, **kwargs):
     """
